+++ tests/fixtures/__init__.py
@@ -1,0 +1,5 @@
+"""Fixtures for tests."""
+
+import os
+
+FIXTURES_PATH = os.path.dirname(__file__)
