+++ tests/fixtures/project_inputs/api_rust_full/systemPatterns.md
