+++ tests/fixtures/project_inputs/api_rust_hobby/codingContext.md
