+++ tests/graph_tests/single_graph/test_architect.py
@@ -27,14 +27,16 @@ TEST_DIR = "./projects/test-site"
 
 
 @pytest.mark.asyncio
-async def test_architect_create_files(pytestconfig):
+async def test_architect_create_files():
     """
     Tests the architect agent graph using langsmith.aevaluate against a LangSmith dataset.
     """
 
     memory_saver = MemorySaver()  # Checkpointer for the graph
     memory_store = InMemoryStore()
-    agent_config = ArchitectConfiguration()
+    agent_config = ArchitectConfiguration(
+        user_id="test_user", model="google_genai:gemini-2.0-flash"
+    )
 
     graph = ArchitectGraph(
         agent_config=agent_config,
@@ -42,23 +44,17 @@ async def test_architect_create_files(pytestconfig):
         store=memory_store,
     )
 
-    config = RunnableConfig(
-        recursion_limit=250,
-        configurable={
-            "thread_id": str(uuid.uuid4()),
-            "user_id": "test_user",
-            "model": "google_genai:gemini-2.0-flash-lite",
-        },
-    )
-
-    # Compile the graph - needs checkpointer for stateful execution during evaluation
-    compiled_graph = graph.builder.compile(
-        checkpointer=memory_saver, store=memory_store
+    config = graph.create_runnable_config(
+        RunnableConfig(
+            recursion_limit=250,
+            configurable={
+                "thread_id": str(uuid.uuid4()),
+            },
+        )
     )
 
     project = Project.from_name("Test Site")
-
-    result = await compiled_graph.ainvoke(
+    result = await graph.compiled_graph.ainvoke(
         State(
             project=project,
             messages=HumanMessage(
