--- tests/integration_tests/__init__.py
@@ -1,1 +1,0 @@
-"""Define any integration tests you want in this directory."""
